+++ src/actions/auth.js
@@ -1,2 +1,0 @@
-export const INIT_SESSION = 'INIT_SESSION';
-export const DESTROY_SESSION = 'DESTROY_SESSION';